Added user input chunks at the very beginning of the script to define directory variables early on so that they could be refered to by consistent variables throughout the script. Started experimenting with chunk options to direct chunks to run or not based on true/false settings by user (work in progress). Revised paths to analysis folder, tools, references (gff), software packages. Adapted script to take in file for metadata already made, and adapted for MiSeq reads. Fixed the broken pipe error when checking raw reads for adater matches with grep prior to using SeqPrep. Added directions and explanations for adapter matching and removal. Fixed adapter sequences to match that were previously not matching to proper location based on seq read structure. Adapted script for merged reads from the MiSeq with options in SeqPrep and removed the use of fastqpairedendvalidator once reads get merged with SeqPrep.
Fixed PrinSeq processing steps that were not doing what we expected for trimming and filtering or generating graphs for the processed files (graphs being generated were for fastq being input for processing rather than graphs of the processed outputs). Made more conservative trimming and filtering options. With the steps repeated for processing, removed poly-a/t tails that remained when only one step of poly-a/t trimming was performed, and allowed better understanding and visualization of filtering and processing effects.
Additional script to analyze level of contamination of non-target reads with ncbi and blast (only tested on one read set).
Cleaned up the script for line length conventions in R script for formating guidelines and removed often repeated lines no longer required to reduce length of script for better readability,
Added functions to clean up output files after running qsubs, for generating PrinSeqGraph files (.gd), and html outputs, gzip and gunzip.
Fixed error for htseq-count by installing htseq in a virtualenv directly in the tools folder (need to see if this is executable when Andre uses it in this directoy).